requirements.txt
aioquic==1.2.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
altgraph==0.17.4
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
Cython==3.0.11
distlib==0.3.9
et-xmlfile==1.1.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.1.0
future==1.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==3.0.2
mitmproxy==11.0.2
mitmproxy-windows==0.10.7
mitmproxy_rs==0.10.7
MouseInfo==0.1.3
msgpack==1.1.0
#netifaces==0.11.0
openpyxl==3.1.5
packaging==24.2
passlib==1.7.4
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
psutil==6.1.0
publicsuffix2==2.20191221
pyarmor==9.0.6
pyarmor.cli.core==7.6.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.21.0
pydivert==2.1.0
PyGetWindow==0.0.9
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
pylsqpack==0.3.18
PyMsgBox==1.0.9
pynput==1.7.7
pyOpenSSL==24.2.1
pyparsing==3.2.0
pyperclip==1.9.0
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
PySide6==6.8.1.1
PySide6_Addons==6.8.1.1
PySide6_Essentials==6.8.1.1
pytest==8.3.3
pytest-assume==2.4.3
pytweening==1.2.0
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
service-identity==24.2.0
shiboken6==6.8.1.1
six==1.16.0
sortedcontainers==2.4.0
tomli==2.1.0
tornado==6.4.2
typing_extensions==4.11.0
urllib3==2.2.3
urwid==2.6.16
virtualenv==20.28.0
wcwidth==0.2.13
Werkzeug==3.1.3
WMI==1.5.1
wsproto==1.2.0
xlwings==0.32.2
zstandard==0.23.0
